✨ Feat: 테마 관련 컴포넌트 및 사이드바 구조 개선, 불필요한 파일 삭제
