+++ src/app/layout.tsx
@@ -1,5 +1,5 @@
 // app/layout.tsx
-import { ThemeProvider } from "@/components/theme-provider";
+import { ThemeProvider } from "@/components/theme/ThemeProvider";
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css"; // 전역 스타일
