--- src/components/app-sidebar.tsx
@@ -1,258 +1,0 @@
-"use client";
-
-import {
-  ChevronsUpDown,
-  Command,
-  FileText,
-  FolderKanban,
-  Home,
-  LayoutDashboard,
-  LogOut,
-  Minus,
-  Settings,
-  UserCircle,
-  type LucideIcon,
-} from "lucide-react";
-import Link from "next/link";
-import { usePathname, useRouter } from "next/navigation";
-import * as React from "react";
-
-// Store 및 Config import
-import { RouteConfig, sidebarRoutes } from "@/config/routes";
-import { AuthState, useAuthStore } from "@/store/authStore";
-
-// Nav 컴포넌트 import
-import { NavMain } from "@/components/nav-main";
-
-// Avatar 컴포넌트 import
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar"; // Avatar 관련 컴포넌트
-// DropdownMenu 컴포넌트 import
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuGroup,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu"; // DropdownMenu 관련 컴포넌트
-import {
-  Sidebar,
-  SidebarContent,
-  SidebarFooter,
-  SidebarHeader,
-  SidebarMenu,
-  SidebarMenuButton,
-  SidebarMenuItem,
-} from "@/components/ui/sidebar";
-
-// 아이콘 매핑
-const iconMap: { [key: string]: LucideIcon } = {
-  Home,
-  FileText,
-  UserCircle,
-  LayoutDashboard,
-  FolderKanban,
-  Settings,
-  Command,
-  LogOut,
-};
-
-// Footer 사용자 표시 컴포넌트
-const UserFooterDisplay = () => {
-  const user = useAuthStore((state: AuthState) => state.user);
-  const router = useRouter();
-
-  if (!user || !user.userRole) {
-    return (
-      <SidebarMenu>
-        <SidebarMenuItem>
-          <SidebarMenuButton className='w-full justify-start' asChild>
-            <Link href='/login'>
-              <UserCircle className='mr-2 size-4' />
-              로그인 필요
-            </Link>
-          </SidebarMenuButton>
-        </SidebarMenuItem>
-      </SidebarMenu>
-    );
-  }
-
-  return (
-    <SidebarMenu>
-      <SidebarMenuItem>
-        <DropdownMenu>
-          <DropdownMenuTrigger asChild>
-            <SidebarMenuButton
-              size='lg'
-              className='w-full data-[state=open]:bg-sidebar-accent data-[state=open]:text-sidebar-accent-foreground'
-            >
-              <Avatar className='mr-2 size-8 rounded-lg'>
-                <AvatarImage
-                  src={(user as any).avatar || undefined}
-                  alt={user.name || "User"}
-                />
-                <AvatarFallback className='rounded-lg'>
-                  {user.name ? user.name.charAt(0).toUpperCase() : "U"}
-                </AvatarFallback>
-              </Avatar>
-              <div className='grid flex-1 text-left text-sm leading-tight'>
-                <span className='truncate font-medium'>
-                  {user.name || "사용자 이름"}
-                </span>
-                <span className='truncate text-xs'>
-                  {user.email || "사용자 이메일"}
-                </span>
-              </div>
-              <ChevronsUpDown className='ml-auto size-4' />
-            </SidebarMenuButton>
-          </DropdownMenuTrigger>
-          <DropdownMenuContent
-            className='w-[calc(var(--radix-dropdown-menu-trigger-width)-8px)] min-w-56 rounded-lg'
-            side='right'
-            align='end'
-            sideOffset={8}
-          >
-            <DropdownMenuLabel className='p-0 font-normal'>
-              <div className='flex items-center gap-2 px-2 py-1.5 text-left text-sm'>
-                <Avatar className='size-8 rounded-lg'>
-                  <AvatarImage
-                    src={(user as any).avatar || undefined}
-                    alt={user.name || "User"}
-                  />
-                  <AvatarFallback className='rounded-lg'>
-                    {user.name ? user.name.charAt(0).toUpperCase() : "U"}
-                  </AvatarFallback>
-                </Avatar>
-                <div className='grid flex-1 text-left text-sm leading-tight'>
-                  <span className='truncate font-medium'>
-                    {user.name || "사용자 이름"}
-                  </span>
-                  <span className='truncate text-xs'>
-                    {user.email || "사용자 이메일"}
-                  </span>
-                </div>
-              </div>
-            </DropdownMenuLabel>
-            <DropdownMenuSeparator />
-            <DropdownMenuGroup>
-              <DropdownMenuItem asChild>
-                <Link href='/mypage'>
-                  <UserCircle className='mr-2 size-4' />
-                  마이페이지
-                </Link>
-              </DropdownMenuItem>
-              <DropdownMenuItem asChild>
-                <Link href='/settings'>
-                  <Settings className='mr-2 size-4' />
-                  설정
-                </Link>
-              </DropdownMenuItem>
-            </DropdownMenuGroup>
-            <DropdownMenuSeparator />
-            <DropdownMenuItem
-              onClick={() => {
-                useAuthStore.getState().logout();
-
-                router.push("/login");
-              }}
-            >
-              <LogOut className='mr-2 size-4' />
-              로그아웃
-            </DropdownMenuItem>
-          </DropdownMenuContent>
-        </DropdownMenu>
-      </SidebarMenuItem>
-    </SidebarMenu>
-  );
-};
-
-// Helper function to convert RouteConfig to NavMain item format
-const convertRoutesToNavItems = (routes: RouteConfig[], pathname: string) => {
-  return routes.map(route => {
-    const IconComponent: LucideIcon | undefined = route.icon
-      ? iconMap[route.icon]
-      : undefined;
-    const resolvedIcon: LucideIcon = IconComponent || Minus; // Fallback icon
-    const isActive = pathname === route.path;
-    return {
-      title: route.name,
-      url: route.path,
-      icon: resolvedIcon,
-      isActive: isActive,
-      disabled: isActive,
-      items: [], // Assuming no sub-items for now
-    };
-  });
-};
-
-export function AppSidebar({ ...props }: React.ComponentProps<typeof Sidebar>) {
-  const userRole = useAuthStore((state: AuthState) => state.user?.userRole);
-  const pathname = usePathname();
-
-  let memberMenuItems: any[] = [];
-  let leaderMenuItems: any[] = [];
-
-  if (userRole) {
-    // userRole이 null이 아닐 때만 메뉴 필터링 실행
-    // Filter MEMBER routes (visible to both MEMBER and LEADER)
-    const memberAccessibleRoutes = sidebarRoutes.filter(
-      // userRole이 null이 아님을 보장하므로 || "" 제거
-      route =>
-        route.viewType === "MEMBER_MENU" && route.roles.includes(userRole)
-    );
-    memberMenuItems = convertRoutesToNavItems(memberAccessibleRoutes, pathname);
-
-    // Filter LEADER-specific routes (visible only to LEADER)
-    if (userRole === "LEADER") {
-      const leaderSpecificRoutes = sidebarRoutes.filter(
-        route =>
-          route.viewType === "LEADER_MENU" && route.roles.includes("LEADER")
-      );
-      leaderMenuItems = convertRoutesToNavItems(leaderSpecificRoutes, pathname);
-    }
-  } else {
-    // userRole이 null이거나 로딩 중일 때 (옵션: 로딩 상태 표시 또는 빈 메뉴)
-    // memberMenuItems와 leaderMenuItems는 이미 빈 배열로 초기화되어 있음
-  }
-
-  const navProjectsItems: any[] = [];
-  const navSecondaryItems: any[] = [];
-
-  return (
-    <Sidebar variant='inset' {...props}>
-      <SidebarHeader>
-        <SidebarMenu>
-          <SidebarMenuItem>
-            <SidebarMenuButton size='lg' asChild>
-              <Link href='/dashboard'>
-                <div className='bg-sidebar-primary text-sidebar-primary-foreground flex aspect-square size-8 items-center justify-center rounded-lg'>
-                  <Command className='size-4' />
-                </div>
-                <div className='grid flex-1 text-left text-sm leading-tight'>
-                  <span className='truncate font-medium'>yAXim</span>
-                  <span className='truncate text-xs'>
-                    사내 주간 보고서 자동 생성 서비스
-                  </span>
-                </div>
-              </Link>
-            </SidebarMenuButton>
-          </SidebarMenuItem>
-        </SidebarMenu>
-      </SidebarHeader>
-
-      <SidebarContent>
-        <NavMain items={memberMenuItems} groupLabel='개인 메뉴' />
-        {userRole === "LEADER" && (
-          <NavMain items={leaderMenuItems} groupLabel='팀장 메뉴' />
-        )}
-        {/* <NavProjects projects={navProjectsItems} /> */}
-        {/* <NavSecondary items={navSecondaryItems} className="mt-auto" /> */}
-      </SidebarContent>
-
-      <SidebarFooter>
-        <UserFooterDisplay />
-      </SidebarFooter>
-    </Sidebar>
-  );
-}
